Add routes to the game and a stats page
This was extremely smooth to implement. When a player enters the
application from "/" as before, they are able to authenticate, create a
game room, and play the game.

When a player enters from a link to a game room, they are able to
authenticate and play as expected.

There will now be a Stats button that pulls you away from the game
to a separate page without causing a refresh as that's how react-router
works. You are still able to get back to the game and this surprisingly
works very smoothly.

Requests to the server for rooms also only happen when you are
redirected to the Authentication component from "/" or "/:room_id".
