+++ src/component/App.tsx
@@ -1,39 +1,18 @@
 import * as React from "react";
-import { useState } from "react";
-import { SignOn } from "./welcome/SignOn";
-import { SlackButton } from "./slack/SlackButton";
-import { Authenticate } from "./slack/Authenticate";
-import * as slack from "./slack/api/handleSlackOAuth";
-import "./css/App.css";
+import { BrowserRouter as Router, Route, Link } from "react-router-dom";
+import { Authentication } from "./Authentication";
+import { Stats } from "./globalstats/Stats";
 
-export const App = () => {
-  const [authenticated, setAuthenticated] = useState<boolean>(false);
+export const App = () => (
+  <Router>
+    <div>
+      <nav>
+        <Link to="/stats">Stats</Link>
+      </nav>
 
-  const slackSignIn = () => {
-    var entryURL = "http://" + window.location.host + window.location.pathname;
-
-    window.location.href =
-      "https://slack.com/oauth/authorize?scope=identity.basic&client_id=708370195111.695122331683&redirect_uri=" +
-      entryURL +
-      "&state=goodtimes";
-  };
-
-  return (
-    <div className="app">
-      <div className="app__title">Welcome to Homerow Brawl</div>
-      <div className="app__bottom">
-        <div className="app__bottom--button">
-          {!sessionStorage.getItem("authenticated") &&
-            window.location.href.length < 45 && (
-              <SlackButton handleEvent={slackSignIn} />
-            )}
-        </div>
-        {!sessionStorage.getItem("authenticated") &&
-          slack.handleURL(setAuthenticated)}
-        {!sessionStorage.getItem("authenticated") &&
-          window.location.href.length > 45 && <Authenticate />}
-        {sessionStorage.getItem("authenticated") && <SignOn />}
-      </div>
+      <Route path="/stats/" component={Stats} />
+      <Route path="/" exact component={Authentication} />
+      <Route path="/:id(\d+)" exact component={Authentication} />
     </div>
-  );
-};
+  </Router>
+);
