+++ src/component/css/Authentication.css
@@ -1,10 +1,10 @@
-.app {
+.authentication {
   margin: 2%;
   border: 2px solid red;
   height: 55em;
 }
 
-.app__title {
+.authentication__title {
   text-align: center;
   font-size: 32pt;
   margin-left: auto;
@@ -14,7 +14,7 @@
   padding: 5%;
 }
 
-.app__bottom {
+.authentication__bottom {
   margin-left: auto;
   margin-right: auto;
   border: 5px solid blue;
@@ -22,6 +22,6 @@
   width: 75%;
 }
 
-.app__bottom--button {
+.authentication__bottom--button {
   text-align: center;
 }
