+++ src/component/globalstats/Stats.tsx
@@ -1,0 +1,2 @@
+import * as React from "react";
+export const Stats = () => <>place stats here</>;
